feat(compression): add zstd compression implementation
