+++ Cargo.toml
@@ -35,6 +35,7 @@ rand = "0.8"
 # performance
 rustc-hash = "1"
 flate2 = "1"
+zstd = "0.13"
 
 [profile.dev]
 opt-level = 1
