+++ msg-wire/Cargo.toml
@@ -18,3 +18,4 @@ thiserror.workspace = true
 tokio-util.workspace = true
 tracing.workspace = true
 flate2.workspace = true
+zstd.workspace = true
