+++ msg-wire/src/compression/gzip.rs
@@ -1,0 +1,53 @@
+use bytes::Bytes;
+use flate2::{read::GzDecoder, write::GzEncoder, Compression};
+use std::io::{self, Read, Write};
+
+use super::{Compressor, Decompressor};
+
+/// A compressor that uses the gzip algorithm.
+pub struct GzipCompressor {
+    level: u32,
+}
+
+impl GzipCompressor {
+    /// Creates a new gzip compressor with the given compression level (0-9).
+    pub fn new(level: u32) -> Self {
+        Self { level }
+    }
+}
+
+impl Compressor for GzipCompressor {
+    fn compress(&self, data: &[u8]) -> Result<Bytes, io::Error> {
+        // Optimistically allocate the compressed buffer to 1/4 of the original size.
+        let mut encoder = GzEncoder::new(
+            Vec::with_capacity(data.len() / 4),
+            Compression::new(self.level),
+        );
+
+        encoder.write_all(data)?;
+
+        let bytes = encoder.finish()?;
+
+        Ok(Bytes::from(bytes))
+    }
+}
+
+#[derive(Debug, Default)]
+pub struct GzipDecompressor;
+
+impl GzipDecompressor {
+    pub fn new() -> Self {
+        Self
+    }
+}
+
+impl Decompressor for GzipDecompressor {
+    fn decompress(&self, data: &[u8]) -> Result<Bytes, io::Error> {
+        let mut decoder = GzDecoder::new(data);
+
+        let mut bytes = Vec::with_capacity(data.len() * 4);
+        decoder.read_to_end(&mut bytes)?;
+
+        Ok(Bytes::from(bytes))
+    }
+}
