+++ msg-wire/src/compression/mod.rs
@@ -1,6 +1,10 @@
 use bytes::Bytes;
-use flate2::{read::GzDecoder, write::GzEncoder, Compression};
-use std::io::{self, Read, Write};
+use std::io;
+
+mod gzip;
+mod zstd;
+pub use gzip::*;
+pub use zstd::*;
 
 /// This trait is used to implement message-level compression algorithms for payloads.
 /// On outgoing messages, the payload is compressed before being sent using the `compress` method.
@@ -16,54 +20,6 @@ pub trait Decompressor: Send + Sync + Unpin + 'static {
     fn decompress(&self, data: &[u8]) -> Result<Bytes, io::Error>;
 }
 
-/// A compressor that uses the gzip algorithm.
-pub struct GzipCompressor {
-    level: u32,
-}
-
-impl GzipCompressor {
-    /// Creates a new gzip compressor with the given compression level (0-9).
-    pub fn new(level: u32) -> Self {
-        Self { level }
-    }
-}
-
-impl Compressor for GzipCompressor {
-    fn compress(&self, data: &[u8]) -> Result<Bytes, io::Error> {
-        // Optimistically allocate the compressed buffer to 1/4 of the original size.
-        let mut encoder = GzEncoder::new(
-            Vec::with_capacity(data.len() / 4),
-            Compression::new(self.level),
-        );
-
-        encoder.write_all(data)?;
-
-        let bytes = encoder.finish()?;
-
-        Ok(Bytes::from(bytes))
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct GzipDecompressor;
-
-impl GzipDecompressor {
-    pub fn new() -> Self {
-        Self
-    }
-}
-
-impl Decompressor for GzipDecompressor {
-    fn decompress(&self, data: &[u8]) -> Result<Bytes, io::Error> {
-        let mut decoder = GzDecoder::new(data);
-
-        let mut bytes = Vec::with_capacity(data.len() * 4);
-        decoder.read_to_end(&mut bytes)?;
-
-        Ok(Bytes::from(bytes))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -72,6 +28,21 @@ mod tests {
     fn test_gzip_compression() {
         let compressor = GzipCompressor::new(6);
         let decompressor = GzipDecompressor::new();
+
+        let data =
+            Bytes::from("hellooooooooooooooooo wwwwwoooooooooooooooooooooooooooooooooooooorld");
+        println!("Before: {:?}", data.len());
+        let compressed = compressor.compress(&data).unwrap();
+        println!("After: {:?}", compressed.len());
+        let decompressed = decompressor.decompress(&compressed).unwrap();
+
+        assert_eq!(data, decompressed);
+    }
+
+    #[test]
+    fn test_zstd_compression() {
+        let compressor = ZstdCompressor::new(6);
+        let decompressor = ZstdDecompressor::new();
 
         let data =
             Bytes::from("hellooooooooooooooooo wwwwwoooooooooooooooooooooooooooooooooooooorld");
